+++ datalad_next/exceptions/__init__.py
@@ -1,6 +1,6 @@
 """All custom exceptions used in datalad-next"""
 
-from datalad.runner import CommandError
+from datalad_next.runners import CommandError
 from datalad.support.exceptions import (
     CapturedException,
     IncompleteResultsError,
