+++ datalad_next/patches/push_to_export_remote.py
@@ -18,7 +18,6 @@ from typing import (
 from unittest.mock import patch
 
 import datalad.core.distributed.push as mod_push
-from datalad.runner.exception import CommandError
 from datalad_next.constraints import EnsureChoice
 from datalad_next.exceptions import CapturedException
 from datalad_next.commands import Parameter
@@ -111,7 +110,7 @@ def get_export_records(repo: AnnexRepo) -> Generator:
             ))
             result_dict["timestamp"] = float(result_dict["timestamp"][:-1])
             yield result_dict
-    except CommandError as command_error:
+    except mod_push.CommandError as command_error:
         # Some errors indicate that there was no export yet.
         # May depend on Git version
         expected_errors = (
@@ -242,7 +241,7 @@ def _transfer_data(repo: AnnexRepo,
                 result_adjusted['action'] = "copy"
                 yield result_adjusted
 
-        except CommandError as cmd_error:
+        except mod_push.CommandError as cmd_error:
             ce = CapturedException(cmd_error)
             yield {
                 **res_kwargs,
